+++ src/vendors/OceanOptics/features/spectrometer/OOISpectrometerFeature.cpp
@@ -282,7 +282,7 @@ long OOISpectrometerFeature::getIntegrationTimeIncrement() const {
 	return this->integrationTimeIncrement;
 }
 
-int OOISpectrometerFeature::getNumberOfPixels() const {
+unsigned short OOISpectrometerFeature::getNumberOfPixels() const {
 	return this->numberOfPixels;
 }
